separate presentational from container component
